remove KDE files
KDE's filemanager Dolphin uses these .directory files,
e.g. to store directory-specific settings such as the ordering.

As I also use dolphin, it will change these files,
and these changes are then detected and proposed to commit by git.

Files starting with a . (hidden files on Linux) also won't be uploaded
to the add-on server, so both wesnoth and git will be in sync here.